Refactoring fixups
