+++ src/lib/route-linestring/index.js
@@ -1,10 +1,10 @@
 import { readFile } from 'node:fs/promises';
 import { resolve, join } from 'node:path';
-import logger from '../lib/logger.js';
+import logger from '../logger.js';
 
 import { lineSlice } from '@turf/line-slice';
 import { lineString, point } from '@turf/helpers';
-import { WEB_APP_GEO_CONFIG_FOLDER_CONTAINER_PATH } from '../config.js';
+import { WEB_APP_GEO_CONFIG_FOLDER_CONTAINER_PATH } from '../../config.js';
 
 let lookupTables = null;
